+++ solution/src/main/java/org/astar/SimpleRecursivePathFinder.java
@@ -1,4 +1,4 @@
-package org.astar.a;
+package org.astar;
 
 
 import java.awt.*;
