+++ solution/src/main/java/org/astar/a/service/AStarPathFinderImpl.java
@@ -1,37 +1,33 @@
-package org.astar.a.c.service;
+package org.astar.a.service;
 
-import org.astar.a.c.value_object.ChainNodeEc;
-import org.astar.a.c.value_object.CoordinateEc;
-import org.astar.a.c.value_object.CoordinateSystemEc;
-import org.eclipse.collections.api.factory.Sets;
-import org.eclipse.collections.api.set.ImmutableSet;
-import org.eclipse.collections.api.set.MutableSet;
-import org.eclipse.collections.impl.list.mutable.FastList;
+import org.astar.a.value_object.ChainNode;
+import org.astar.a.value_object.Coordinate;
+import org.astar.a.value_object.CoordinateSystem;
+import org.astar.api.AStarPathFinder;
 
-import java.util.Map;
-import java.util.Objects;
-import java.util.Optional;
+import java.util.*;
 import java.util.function.Predicate;
+import java.util.stream.Stream;
 
 /**
  * A star path finder.
  */
-public class AStarPathFinderEcImpl {
+public class AStarPathFinderImpl implements AStarPathFinder {
 
     /**
      * 座標系
      */
-    private CoordinateSystemEc coordinateSystemEc;
+    private CoordinateSystem coordinateSystem;
 
     /**
      * 暫定的に探索済みの最短経路で無いと判定された経路(Node)を格納したリスト
      */
-    private MutableSet<ChainNodeEc> provisionalOpenSet = Sets.mutable.of();
+    private Set<ChainNode> provisionalOpenSet = new HashSet<>();
 
     /**
      * 暫定的に最短経路であると評価された経路(Node)を格納したリスト
      */
-    private MutableSet<ChainNodeEc> provisionalCloseSet = Sets.mutable.of();
+    private Set<ChainNode> provisionalCloseSet = new HashSet<>();
 
     /**
      * 起点から終点までの最短経路を探索する
@@ -40,27 +36,28 @@ public class AStarPathFinderEcImpl {
      * 経路が存在しない場合nullを返却.
      * </p>
      *
-     * @param width             　最大幅
-     * @param height            　最大高さ
-     * @param startCoordinateEc 　起点
-     * @param goalCoordinateEc  　終点
-     * @param obstacles         　障害物の位置の一覧
+     * @param width           　最大幅
+     * @param height          　最大高さ
+     * @param startCoordinate 　起点
+     * @param goalCoordinate  　終点
+     * @param obstacles       　障害物の位置の一覧
      * @return 起点から終点までの経路
      */
-    public Optional<FastList<CoordinateEc>> findPath(
+    @Override
+    public Optional<List<Coordinate>> findPath(
             int width,
             int height,
-            CoordinateEc startCoordinateEc,
-            CoordinateEc goalCoordinateEc,
-            Optional<ImmutableSet<CoordinateEc>> obstacles) {
+            Coordinate startCoordinate,
+            Coordinate goalCoordinate,
+            Optional<Set<Coordinate>> obstacles) {
 
-        ImmutableSet<CoordinateEc> concreteObstacles = obstacles.orElseGet(Sets.immutable::empty);
-        this.coordinateSystemEc = new CoordinateSystemEc(width, height, startCoordinateEc, goalCoordinateEc, concreteObstacles);
+        Set<Coordinate> concreteObstacles = obstacles.orElseGet(Collections::emptySet);
+        this.coordinateSystem = new CoordinateSystem(width, height, startCoordinate, goalCoordinate, concreteObstacles);
 
         //発散
-        diverge(startCoordinateEc, Optional.empty());
+        diverge(startCoordinate, Optional.empty());
 
-        FastList<CoordinateEc> result = findPath();
+        List<Coordinate> result = findPath();
         return Optional.ofNullable(result);
     }
 
@@ -69,9 +66,9 @@ public class AStarPathFinderEcImpl {
      *
      * @return
      */
-    private FastList<CoordinateEc> findPath() {
+    private List<Coordinate> findPath() {
 
-        ChainNodeEc shortestNode = null;
+        ChainNode shortestNode = null;
         while (isContinue()) {
             //最短経路の収束
             shortestNode = converge();
@@ -82,14 +79,14 @@ public class AStarPathFinderEcImpl {
         if (isUnreachableGoal(shortestNode)) {
             return null;
         } else {
-            var path = new FastList<CoordinateEc>();
-            path.add(shortestNode.getCoordinateEc());
+            var path = new ArrayList<Coordinate>();
+            path.add(shortestNode.getCoordinate());
             while (shortestNode.getPreviousNode().isPresent()) {
                 shortestNode = shortestNode.getPreviousNode().get();
-                path.add(shortestNode.getCoordinateEc());
+                path.add(shortestNode.getCoordinate());
             }
-
-            return path.reverseThis();
+            Collections.reverse(path);
+            return path;
         }
 
     }
@@ -102,21 +99,21 @@ public class AStarPathFinderEcImpl {
      */
     private boolean isContinue() {
         return this.provisionalOpenSet.size() != 0
-                && !this.provisionalCloseSet
-                .anySatisfy(node -> this.coordinateSystemEc.isGoal(node.getCoordinateEc()));
+                && !this.provisionalCloseSet.stream()
+                .anyMatch(node -> this.coordinateSystem.isGoal(node.getCoordinate()));
     }
 
     /**
      * 目標地点に到達しないか？
      *
-     * @param chainNodeEc
+     * @param chainNode
      * @return true:到達しない場合
      * false:到達する場合
      */
-    private boolean isUnreachableGoal(ChainNodeEc chainNodeEc) {
-        return Objects.isNull(chainNodeEc)
-                || !this.provisionalCloseSet
-                .anySatisfy(node -> this.coordinateSystemEc.isGoal(node.getCoordinateEc()));
+    private boolean isUnreachableGoal(ChainNode chainNode) {
+        return Objects.isNull(chainNode)
+                || !this.provisionalCloseSet.stream()
+                .anyMatch(node -> this.coordinateSystem.isGoal(node.getCoordinate()));
     }
 
     /**
@@ -124,9 +121,9 @@ public class AStarPathFinderEcImpl {
      *
      * @return 推定最短経路を保持するノード
      */
-    private ChainNodeEc converge() {
+    private ChainNode converge() {
         //最小ノードを取得
-        ChainNodeEc shortestNode = findShortestChainNodeInOpenList();
+        ChainNode shortestNode = findShortestChainNodeInOpenList();
         this.provisionalCloseSet.add(shortestNode);
         this.provisionalOpenSet.remove(shortestNode);
         return shortestNode;
@@ -137,9 +134,10 @@ public class AStarPathFinderEcImpl {
      *
      * @return 最短経路を返却
      */
-    private ChainNodeEc findShortestChainNodeInOpenList() {
-        return this.provisionalOpenSet
-                .min(ChainNodeEc::compareTo);
+    private ChainNode findShortestChainNodeInOpenList() {
+        return this.provisionalOpenSet.stream()
+                .min(ChainNode::compareTo)
+                .get();
     }
 
     /**
@@ -147,16 +145,16 @@ public class AStarPathFinderEcImpl {
      *
      * @param centerNode 発散時に中央に存在する座標を持つNode
      */
-    private void divergeAroundWith(ChainNodeEc centerNode) {
-        final CoordinateEc center = centerNode.getCoordinateEc();
+    private void divergeAroundWith(ChainNode centerNode) {
+        final Coordinate center = centerNode.getCoordinate();
 
         final var previousNode = Optional.of(centerNode);
 
-        final var aroundPoints = FastList.newListWith(
-                Map.entry(center.left(), this.coordinateSystemEc.isValidLeft()),
-                Map.entry(center.up(), this.coordinateSystemEc.isValidUp()),
-                Map.entry(center.right(), this.coordinateSystemEc.isValidRight()),
-                Map.entry(center.down(), this.coordinateSystemEc.isValidDown())
+        final var aroundPoints = Stream.of(
+                Map.entry(center.left(), this.coordinateSystem.isValidLeft()),
+                Map.entry(center.up(), this.coordinateSystem.isValidUp()),
+                Map.entry(center.right(), this.coordinateSystem.isValidRight()),
+                Map.entry(center.down(), this.coordinateSystem.isValidDown())
         );
         aroundPoints.forEach(e -> divergeIf(e.getKey(), previousNode, e.getValue()));
 
@@ -165,25 +163,25 @@ public class AStarPathFinderEcImpl {
     /**
      * 妥当性のある座標の場合、経路の発散処理を行う
      *
-     * @param coordinateEc      座標
+     * @param coordinate        座標
      * @param previousNode      一つ前のNode
      * @param isValidCoordinate 座標の妥当性を判定する関数
      */
-    private void divergeIf(CoordinateEc coordinateEc, Optional<ChainNodeEc> previousNode, Predicate<CoordinateEc> isValidCoordinate) {
-        if (isValidCoordinate.test(coordinateEc)) {
-            diverge(coordinateEc, previousNode);
+    private void divergeIf(Coordinate coordinate, Optional<ChainNode> previousNode, Predicate<Coordinate> isValidCoordinate) {
+        if (isValidCoordinate.test(coordinate)) {
+            diverge(coordinate, previousNode);
         }
     }
 
     /**
      * 新規Node(経路)のコストが新規または、探索済みの同じ座標を持つNode(経路)より総コストが低い場合、経路をOpenSetに追加し、経路を発散させる。
      *
-     * @param processCoordinateEc 処理中の座標
-     * @param previousNode        一つ前のNode
+     * @param processCoordinate 処理中の座標
+     * @param previousNode      一つ前のNode
      */
-    private void diverge(CoordinateEc processCoordinateEc, Optional<ChainNodeEc> previousNode) {
+    private void diverge(Coordinate processCoordinate, Optional<ChainNode> previousNode) {
         //新しいNodeのコスト計算
-        ChainNodeEc newNode = newNode(processCoordinateEc, previousNode);
+        ChainNode newNode = newNode(processCoordinate, previousNode);
         //計算済みのNodeの方より新しくコスト算出したノードの方がコストが低いか判定
         if (absentsAfterRemoveIfInCloseList(newNode)) {
             //新しいNodeをOpenリストに追加
@@ -198,8 +196,8 @@ public class AStarPathFinderEcImpl {
      * @param previousNode       一つ前のNode
      * @return 重みづけされた新規Node
      */
-    private ChainNodeEc newNode(CoordinateEc currentCoordinates, Optional<ChainNodeEc> previousNode) {
-        return ChainNodeEc.newNode(currentCoordinates, previousNode, this.coordinateSystemEc);
+    private ChainNode newNode(Coordinate currentCoordinates, Optional<ChainNode> previousNode) {
+        return ChainNode.newNode(currentCoordinates, previousNode, this.coordinateSystem);
     }
 
     /**
@@ -216,13 +214,14 @@ public class AStarPathFinderEcImpl {
      * false:新規Nodeのheadの座標が暫定最短経路一覧内のheadの座標と同じ場合に既存Nodeの総コストが低い場合、既存のNodeを一覧から削除しない。
      * </p>
      */
-    private boolean absentsAfterRemoveIfInCloseList(ChainNodeEc newNode) {
+    private boolean absentsAfterRemoveIfInCloseList(ChainNode newNode) {
 
         //新規Nodeと同じ座標の計算済みNodeの最短経路のコストが新規Nodeより大きい場合リストから削除
-        this.provisionalCloseSet
-                .select(node -> node.equalsCoordinate(newNode))
-                .removeIf(node -> newNode.lessThanTotalCost(node));
-
+        this.provisionalCloseSet.stream()
+                .filter(node -> node.equalsCoordinate(newNode))
+                .filter(node -> newNode.lessThanTotalCost(node))
+                .findFirst()
+                .ifPresent(oldNode -> this.provisionalCloseSet.remove(oldNode));
         //計算済みNode内の方が最短経路のコストが小さい場合
         if (this.provisionalCloseSet.contains(newNode)) {
             return false;
@@ -241,9 +240,10 @@ public class AStarPathFinderEcImpl {
      *
      * @param newNode 新規Node
      */
-    private void addOrUpdateInOpenList(ChainNodeEc newNode) {
-        this.provisionalOpenSet
-                .detectOptional(node -> node.equalsCoordinate(newNode))
+    private void addOrUpdateInOpenList(ChainNode newNode) {
+        this.provisionalOpenSet.stream()
+                .filter(node -> node.equalsCoordinate(newNode))
+                .findFirst()
                 .ifPresentOrElse(
                         oldNode -> {
                             if (newNode.lessThanTotalCost(oldNode)) {
@@ -253,4 +253,5 @@ public class AStarPathFinderEcImpl {
                         () -> this.provisionalOpenSet.add(newNode)
                 );
     }
+
 }
