+++ solution/src/main/java/org/astar/a/value_object/ChainNode.java
@@ -1,4 +1,4 @@
-package org.astar.a.a.value_object;
+package org.astar.a.value_object;
 
 import lombok.Getter;
