+++ solution/src/main/java/org/astar/b/value_object/ChainNode_b.java
@@ -1,4 +1,4 @@
-package org.astar.a.b.value_object;
+package org.astar.b.value_object;
 
 import lombok.Getter;
